Changed: refactored hardhat.config.js and error message thrown when user denies wanting to use plain-text secrets in .env file
